Support stop/parity bits on serial port
bug #23952

Unfortunately, screen's handling of serial ports is
somewhat incomplete: You can set the baud rate and the
word length (cs7), but not the number of stop bits
and/or the parity settings.

Since the OS calls are readily available, adding this
to a future version of screen would proably be easy
enough to do inside tty.sh.